Use --with-opt-dir instead of --with-xx-dir that ruby isn't respecting. Pass from conanfile.py to cmake
conanio/gccXX (eg 10) removed the libxxx-dev (eg libgmp-dev) from the image. This made me realize that the conan deps weren't being picked up by the build. The openssl was, because --with-openssl-dir is explicitly used.

So here, we rely on --with-opt-dir.

I would like to get rid of CMakeLists / build_ruby.bat.in altogether, but this isn't worth it until we move to ruby 3.x so I'm just ductaping it here.